Initialise vector, node, main and const
Created two module, node contain the Node class and vector contain the Vector2 class, const will contain all constants in the future.

The Vector2 class useful for position in 2D space and storing two variable.
The Node class that will serve as a base to make a tree structure, base for Node with two dimensional position, and to make the update() and draw() function of object way simpler.